insert, constructors rewritten
our hashtable array initially stored elements, but when we inserted new elements and tried to point their next attributes to the head in the hashtable, the address retrieved was the index of the head in the implementation array. effectively, inserting an element makes its next attribute point to itself (or did).